requirements.txt
aiohttp==0.21.6
aiohttp-jinja2==0.7.0
chardet==2.3.0
Jinja2==2.8
MarkupSafe==0.23
msgpack-python==0.4.7
